create register and LoginPage componensts
