+++ src/components/LoginPage/LoginPage.js
@@ -1,0 +1,11 @@
+import Login from "./Login";
+import Register from "./Register";
+const LoginPage = () => {
+    return (
+<section>
+    <Login />
+    <Register />
+</section>
+    )
+}
+export default LoginPage;
